backend/requirements.txt
fastapi
pydantic
uvicorn[standard]
python-multipart
openai
pypdf
python-dateutil